device: manage peer.queue.outbound locally
The outbound queue has exactly one writer and one reader.
Switch it from a struct field to a local variable,
and enforce readonly/writeonly with the type system.

The writer is responsible for closing the channel
when it is done writing to it.

Because we know that the channel will be closed
when all writes are completed, we can simplify
draining the channel to a range loop.

While we're here, switch some comments to a more idiomatic godoc style.

Signed-off-by: Josh Bleecher Snyder <josh@tailscale.com>