device: eliminate the select in RoutineSequentialSender
When peer.routines.stop is closed, RoutineNonce will close the outboundQueue.
RoutineSequentialSender thus does not need to monitor peer.routines.stop;
it suffices to read until outboundQueue is closed.

We can eliminate the select statement, which is costly (in a high
performance, latency sensitive setting), in favor of ranging over outboundQueue.

Because we are now ranging over every element of outboundQueue
in the main body, we no longer need a separate drain step
on function exit.

Signed-off-by: Josh Bleecher Snyder <josh@tailscale.com>